finalize the project
I make some small change to the menu and import the decoding function
and its related code from Paul, as we decided to make the decode tree
based on the text file.
I tested these codes both for encoding and decoding, and deleted any
other codes that are not adaptive to the completed version. So this is
the final version of the project.